VrepRemoveCollisionMesh
