+++ src/compas_fab/backends/vrep/backend_features/__init__.py
@@ -1,8 +1,9 @@
 from .vrep_add_attached_collision_mesh import VrepAddAttachedCollisionMesh
 from .vrep_add_collision_mesh import VrepAddCollisionMesh
 from .vrep_forward_kinematics import VrepForwardKinematics
-from .vrep_plan_motion import VrepPlanMotion
 from .vrep_inverse_kinematics import VrepInverseKinematics
+from .vrep_plan_motion import VrepPlanMotion
+from .vrep_remove_collision_mesh import VrepRemoveCollisionMesh
 
 __all__ = [
     'VrepForwardKinematics',
@@ -10,4 +11,5 @@ __all__ = [
     'VrepPlanMotion',
     'VrepAddAttachedCollisionMesh',
     'VrepAddCollisionMesh',
+    'VrepRemoveCollisionMesh',
 ]
