+++ src/compas_fab/backends/vrep/planner_backend_vrep.py
@@ -8,6 +8,7 @@ from compas_fab.backends.vrep.backend_features.vrep_add_collision_mesh import Vr
 from compas_fab.backends.vrep.backend_features.vrep_forward_kinematics import VrepForwardKinematics
 from compas_fab.backends.vrep.backend_features.vrep_inverse_kinematics import VrepInverseKinematics
 from compas_fab.backends.vrep.backend_features.vrep_plan_motion import VrepPlanMotion
+from compas_fab.backends.vrep.backend_features.vrep_remove_collision_mesh import VrepRemoveCollisionMesh
 
 
 class VrepPlanner(PlannerInterface):
@@ -20,3 +21,4 @@ class VrepPlanner(PlannerInterface):
         self.add_attached_collision_mesh = VrepAddAttachedCollisionMesh(self.client)
         self.pick_building_member = VrepAddAttachedCollisionMesh(self.client).pick_building_member
         self.add_collision_mesh = VrepAddCollisionMesh(self.client)
+        self.remove_collision_mesh = VrepRemoveCollisionMesh(self.client)
